Make it possible to test different nodes with different hieradata
The govuk_nodes_spec_optional.rb spec now tests each node with the
Hieradata for each environment, and can exclude nodes that are not
found in an environment. I'm not sure how to do this exclusion
automatically, so it's just a array of class names in the test at the
moment.

This change should help resolve problems I've been having recently
where I've been changing the Puppet configuration relating to AWS, but
the tests have expected me to have that configuration work in Carrenza.