Fix load balancer healthcheck for whitehall-admin.
Plumb the healthcheck for `whitehall-admin` through the default vhost in
the nginx config.

The healthcheck needs to be accessible to the load balancer so that
traffic does not get routed to broken instances. The convention for the
URLs paths of these healthchecks is defined in
https://github.com/alphagov/govuk-aws/blob/master/doc/architecture/decisions/0037-alb-health-checks.md

`whitehall-admin` needs some special-case config to achieve this because
it sets `enable_nginx_vhost=false` and does the nginx host config
itself, which means we don't get this for free from
`govuk::app::config`.