Add Icinga checks for email-alert-api messages
This adds Icinga checks that monitor the number of unprocessed
messages in the `email-alert-api` that have a `crtical` and
`warning` latency (created more then `x` minutes ago but have
not been processed yet). They are not machine specific and the
data used lived in `stats/gauges/govuk/email-alert-api` in Graphite.

We will be alerted when either values are greater than `0`.