Support running only app-specific scenarios
This adds an optional string parameter that will be used by the
'jenkins.sh' script in Smokey to only run scenarios with a tag
that matches, noting that many scenarios are not relevant to a
given app. We need this feature to support Continuous Deployment,
so that deploys are not blocked by unrelated failures. The name
'TARGET_APPLICATION' makes this job triggerable by 'Deploy_App',
since the parameter names match.