Fix missing release app token in CI
Previously we added an HTTP API call to the production release app
to abort downstream deploys if they are meant to be frozen. These
jobs are failing in CI because the environment variable used to get
the bearer token is missing. It's unclear where it comes from.

This ensures the variable is explicitly passed to the job, and moves
it and the other GitHub secret into a wrapper, so they are not exposed
in the console output of the job.