Update Notify Emails Sent (per 30s)
This updates the dashboard to be more accurate as the rate limit
previously was 360r/s (10800 / 30) whereas it's actually 350r/s so the
limit line has been updated.

Have also updated the wording from a minute to 30s as the numbers being
displayed did not correlate with the limit over a minute, rather the
limit over 30 seconds.