Remove undocumented unicorn worker check
This check alerts much too frequently due to app restarts, which
can last for several minutes. It has no associated documentation,
indicating it has little value. The alert was created in 2013, but
with no explanation [1]. Since we have other alerts to tell us if
an app is running, we can afford to lose this one.

[1]: effbd06dffa52ef9ef5e97a1ab6a51e107d914b4