DRY up the data sync start/end times
These values are beginning to leak into multiple places. It is good
maintenance practice to try to keep them in one place.

There is probably a cleaner way of doing this with YML config, but
given that some places need the start/end times and others need
specifically the start/end hour and minute, a function seems a
reasonable thing to create to encapsulate all that knowledge.

In any case, it is an improvement on trying to duplicate the
values throughout the repository.