Add frontend memcache servers to production
We've completed a round of testing on the memcache configuration and
concluded that we think it's beneficial to apply this. I've applied this
by setting this in the common file and then overriding it in draft with
the node itself. This seemed the most consistent approach.

In testing we confirmed that this shared memcache instance dropped
response times by around 50% compared to having to request a postcode
from mapit.

We also found that when hit with a load that was enough to overwhelm our
app instances we only seemed to utilise about 1% of the CPU and
available memory was still very high. Finally we tested how Rails copes
with a failure and this seemed to be graceful, with an application just
executing the non-cached code if the resource is unavailable.